requirements.txt
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
cssselect==1.1.0
idna==2.10
iniconfig==1.1.1
lxml==4.6.3
packaging==20.9
pluggy==0.13.1
py==1.10.0
pymongo==3.11.4
pyparsing==2.4.7
PySimpleGUI==4.41.2
pytest==6.2.4
python-dotenv==0.17.1
requests==2.25.1
toml==0.10.2
urllib3==1.26.4
